requirements.txt
# Static Analysis
pre-commit
# Configuration
python-dotenv
pydantic-settings
# API
fastapi
uvicorn
# DB
pymongo
# Test
pytest
httpx
# Profile
memory_profiler
line_profiler
matplotlib
# boto3
boto3_type_annotations==0.3.1
boto3==1.29.3
absl-py==2.0.0
argument==1.4.0
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.2
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.0.4
executing==0.8.3
filelock==3.12.4
flatbuffers==1.12
fsspec==2023.9.1
gast==0.4.0
google-auth==2.23.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.58.0
h5py==3.9.0
huggingface-hub==0.17.2
idna==3.4
ipykernel==6.25.0
ipython==8.15.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.2
JPype1==1.4.1
jupyter_client==8.1.0
jupyter_core==5.3.0
keras==2.9.0
Keras-Preprocessing==1.1.2
konlpy==0.6.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
numpy==1.26.0
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.36
protobuf==3.19.6
psutil==5.9.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pykospacing==0.5
python-dateutil==2.8.2
pywin32==305.1
PyYAML==6.0.1
pyzmq==25.1.0
rank-bm25==0.2.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==68.0.0
six==1.16.0
stack-data==0.2.0
sympy==1.12
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.3
tensorflow-estimator==2.9.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
traitlets==5.7.1
transformers==4.33.2
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.5
Werkzeug==2.3.7
wheel==0.38.4
wrapt==1.14.1
